Fixed connection b/w frontend and backend for fetch
Switching the target link from localhost:5000 to 127.0.0.1:5000
fixed any fetch requests using absolute paths. However, for
relative paths, there was an issue with the syntax of the
proxy in vite.config.ts. Removing the object after '/api'
and just putting in a string of the target link solved
these issues, so now the file can be uploaded to the
flask server, and the fileId is successfully returned to be
stored in local storage.