Build model and insights and store in global cache
The functions defined within the files in the model folder
are used to build the model based on the csv from the
fileId. The model and data are used to generate the
insights, equivalence insights, and shap values to be
fetched by the front end from the other routes to
be defined in the backend later.